+++ Assets/Xsolla.Demo/Store/Scripts/Sdk/StoreLogic.cs
@@ -1,3 +1,4 @@
+using UnityEngine;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -22,9 +23,10 @@ namespace Xsolla.Demo
 		public static void PurchaseForRealMoney(CatalogItemModel item, Action<CatalogItemModel> onSuccess, Action<Error> onError)
 		{
 			XsollaCatalog.Purchase(
-				item.Sku,
-				_ => CompletePurchase(item, () => onSuccess?.Invoke(item)),
-				error => OnPurchaseError(error, onError),
+				itemSku: item.Sku,
+				onSuccess: _ => CompletePurchase(item, () => onSuccess?.Invoke(item)),
+				onError: error => OnPurchaseError(error, onError),
+				onBrowseClosed: _ => Debug.Log("Payment browser closed"),
 				customHeaders: GenerateCustomHeaders());
 		}
 
@@ -63,7 +65,7 @@ namespace Xsolla.Demo
 		{
 			if (!items.Any())
 			{
-				var error = new Error(errorMessage: "Cart is empty");
+				var error = new Error(ErrorType.InvalidData, errorMessage: "Cart is empty");
 				onError?.Invoke(error);
 				return;
 			}
@@ -103,7 +105,7 @@ namespace Xsolla.Demo
 		{
 			if (!items.Any())
 			{
-				var error = new Error(errorMessage: "Cart is empty");
+				var error = new Error(ErrorType.InvalidData, errorMessage: "Cart is empty");
 				onError?.Invoke(error);
 				return;
 			}